requirements.txt
# APE Production Requirements
# Core dependencies only - no testing/development tools

# Core Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # For Parquet support

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0  # For interactive visualizations including Sankey diagrams

# Configuration
pyyaml>=6.0

# UI Components
streamlit-carbon-button>=1.0.0  # Carbon Design System buttons

# Utilities
haikunator>=2.1.0  # For memorable simulation names
pendulum>=3.0.0  # For human-friendly datetime formatting
psutil>=5.9.0  # For memory monitoring